payllm-server/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
docstring_parser==0.16
fastapi==0.103.1
google-api-core==2.24.2
google-auth==2.40.1
google-cloud-aiplatform==1.92.0
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.10.0
google-crc32c==1.7.1
google-genai==1.14.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
numpy==2.2.5
packaging==25.0
proto-plus==1.26.1
protobuf==4.25.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.3.0
pydantic_core==2.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9.1
shapely==2.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.23.2
websockets==15.0.1
